PhpCode: Handle empty constant
When generating a constant name for a string, we were assuming that we'd
always be able to access `$phpName[0]`, but this isn't the case when
there are no "usable" characters in the original string (e.g., `0`, `_`,
`=`), resulting in an empty transformed string.

Rearrange the logic to make sure that we don't try to access the first
character of an empty string. This bug was surfaced by trying to create
an enum constant for the string `=`.